Added pH sensor readings, and average temperature readings
Added pH sensor reading function (GetPHReading()) using the built in ADC of the micro controller. It works by a simple sampling process, which ends when a flag is set by the pH sensor. Then the pH value is read, and LEDs are turned on based on that. Also added functions to get average readings (AverageReadings) which stores how many conversions have been made in a cycle, and finds the mean. Makes LED blink based on the information.